removed unused mobile menu, removed unused transcript js script
