+++ app/src/pages/login.html
@@ -1,30 +1,44 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>GPA Calculator - Login</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;600&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="../style/style.css">
-</head>
-<body class="d-flex align-items-center justify-content-center vh-100">
-    <div class="card p-4 shadow" style="width: 400px;">
-        <h3 class="text-center mb-4">Student Login</h3>
-        <form id="loginForm">
-            <div class="mb-3">
-                <label class="form-label">University Email</label>
-                <input type="email" id="email" class="form-control" placeholder="name@univ.edu" required>
-            </div>
-            <div class="mb-3">
-                <label class="form-label">Password</label>
-                <input type="password" id="password" class="form-control" required>
-            </div>
-            <button type="submit" class="btn btn-primary w-100">Login</button>
-        </form>
-        <p class="text-center mt-3">Don't have an account? <a href="register.html">Register here</a></p>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+    />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;600&display=swap"
+      rel="stylesheet"
+    />
+    <link rel="stylesheet" href="../style/style.css" />
+  </head>
+  <body class="d-flex align-items-center justify-content-center vh-100">
+    <div class="card p-4 shadow" style="width: 400px">
+      <h3 class="text-center mb-4">Student Login</h3>
+      <form id="loginForm">
+        <div class="mb-3">
+          <label class="form-label">Username</label>
+          <input
+            type="text"
+            id="username"
+            class="form-control"
+            placeholder="username"
+            required
+          />
+        </div>
+        <div class="mb-3">
+          <label class="form-label">Password</label>
+          <input type="password" id="password" class="form-control" required />
+        </div>
+        <button type="submit" class="btn btn-primary w-100">Login</button>
+      </form>
+      <p class="text-center mt-3">
+        Don't have an account? <a href="register.html">Register here</a>
+      </p>
     </div>
     <script src="../scripts/auth.js" type="module"></script>
-<script src="../scripts/theme.js"></script>
-</body>
+    <script src="../scripts/theme.js"></script>
+  </body>
 </html>
