+++ app/src/pages/mark_scaling.html
@@ -1,71 +1,113 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Grade Scaling Reference - Student Success Hub</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="../style/style.css">
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&display=swap" rel="stylesheet">
-</head>
-<body>
-            <div class="sidebar">
-        <div class="sidebar-header">🎓 Success Hub</div>
-        <ul class="sidebar-menu">
-            <li><a href="dashboard.html">📊 Dashboard</a></li>
-            <li><a href="semesters.html">📅 Semesters</a></li>
-            <li><a href="courses.html">📚 Courses</a></li>
-            <li><a href="calculator.html">🧮 GPA Calculator</a></li>
-            <li><a href="simulator.html">🎯 Goal Simulator</a></li>
-            <li><a href="mark_scaling.html" class="active">📈 Mark Scaling</a></li>
-            <li><a href="transcript.html">📜 Transcript</a></li>
-            <li><a href="settings.html">👤 Profile Settings</a></li>
-            <li class="mt-auto"><a href="#" id="logoutBtn" style="color: #f87171;">🚪 Logout</a></li>
-        </ul>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+    />
+    <link rel="stylesheet" href="../style/style.css" />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&display=swap"
+      rel="stylesheet"
+    />
+  </head>
+  <body>
+    <div class="sidebar">
+      <div class="sidebar-header">🎓 Success Hub</div>
+      <ul class="sidebar-menu">
+        <li><a href="dashboard.html">📊 Dashboard</a></li>
+        <li><a href="semesters.html">📅 Semesters</a></li>
+        <li><a href="courses.html">📚 Courses</a></li>
+        <li><a href="calculator.html">🧮 GPA Calculator</a></li>
+        <li><a href="simulator.html">🎯 Goal Simulator</a></li>
+        <li><a href="mark_scaling.html" class="active">📈 Mark Scaling</a></li>
+        <li><a href="transcript.html">📜 Transcript</a></li>
+        <li><a href="settings.html">👤 Profile Settings</a></li>
+        <li class="mt-auto">
+          <a href="#" id="logoutBtn" style="color: #f87171">🚪 Logout</a>
+        </li>
+      </ul>
     </div>
 
-    <button id="mobile-toggle" class="mobile-toggle" aria-label="Toggle Menu">
-        ☰
-    </button>
-
     <div class="main-content">
-        <div class="mb-4">
-            <h2 class="fw-bold">Grade Scaling Reference</h2>
-            <p class="text-muted">The standard conversion table used for calculating GPA points.</p>
-        </div>
-        <div class="row justify-content-center">
-            <div class="col-md-6">
-                <div class="custom-card">
-                    <div class="table-responsive">
-                        <table class="table table-hover text-center align-middle">
-                            <thead class="table-light">
-                                <tr>
-                                    <th>Letter Grade</th>
-                                    <th>GPA Points</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                <tr><td><strong>A</strong></td><td>4.0</td></tr>
-                                <tr><td><strong>A-</strong></td><td>3.7</td></tr>
-                                <tr><td><strong>B+</strong></td><td>3.3</td></tr>
-                                <tr><td><strong>B</strong></td><td>3.0</td></tr>
-                                <tr><td><strong>B-</strong></td><td>2.7</td></tr>
-                                <tr><td><strong>C+</strong></td><td>2.3</td></tr>
-                                <tr><td><strong>C</strong></td><td>2.0</td></tr>
-                                <tr><td><strong>C-</strong></td><td>1.7</td></tr>
-                                <tr><td><strong>D+</strong></td><td>1.3</td></tr>
-                                <tr><td><strong>D</strong></td><td>1.0</td></tr>
-                                <tr><td><strong>D-</strong></td><td>0.7</td></tr>
-                                <tr><td><strong>F</strong></td><td>0.0</td></tr>
-                            </tbody>
-                        </table>
-                    </div>
-                </div>
+      <div class="mb-4">
+        <h2 class="fw-bold">Grade Scaling Reference</h2>
+        <p class="text-muted">
+          The standard conversion table used for calculating GPA points.
+        </p>
+      </div>
+      <div class="row justify-content-center">
+        <div class="col-md-6">
+          <div class="custom-card">
+            <div class="table-responsive">
+              <table class="table table-hover text-center align-middle">
+                <thead class="table-light">
+                  <tr>
+                    <th>Letter Grade</th>
+                    <th>GPA Points</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  <tr>
+                    <td><strong>A</strong></td>
+                    <td>4.0</td>
+                  </tr>
+                  <tr>
+                    <td><strong>A-</strong></td>
+                    <td>3.7</td>
+                  </tr>
+                  <tr>
+                    <td><strong>B+</strong></td>
+                    <td>3.3</td>
+                  </tr>
+                  <tr>
+                    <td><strong>B</strong></td>
+                    <td>3.0</td>
+                  </tr>
+                  <tr>
+                    <td><strong>B-</strong></td>
+                    <td>2.7</td>
+                  </tr>
+                  <tr>
+                    <td><strong>C+</strong></td>
+                    <td>2.3</td>
+                  </tr>
+                  <tr>
+                    <td><strong>C</strong></td>
+                    <td>2.0</td>
+                  </tr>
+                  <tr>
+                    <td><strong>C-</strong></td>
+                    <td>1.7</td>
+                  </tr>
+                  <tr>
+                    <td><strong>D+</strong></td>
+                    <td>1.3</td>
+                  </tr>
+                  <tr>
+                    <td><strong>D</strong></td>
+                    <td>1.0</td>
+                  </tr>
+                  <tr>
+                    <td><strong>D-</strong></td>
+                    <td>0.7</td>
+                  </tr>
+                  <tr>
+                    <td><strong>F</strong></td>
+                    <td>0.0</td>
+                  </tr>
+                </tbody>
+              </table>
             </div>
+          </div>
         </div>
+      </div>
     </div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     <script src="../scripts/ui.js"></script>
     <script src="../scripts/theme.js"></script>
-</body>
+  </body>
 </html>
