+++ app/src/pages/semesters.html
@@ -1,79 +1,84 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>My Semesters - Student Success Hub</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="../style/style.css">
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&display=swap" rel="stylesheet">
-</head>
-<body>
-
-            <div class="sidebar">
-        <div class="sidebar-header">🎓 Success Hub</div>
-        <ul class="sidebar-menu">
-            <li><a href="dashboard.html">📊 Dashboard</a></li>
-            <li><a href="semesters.html" class="active">📅 Semesters</a></li>
-            <li><a href="courses.html">📚 Courses</a></li>
-            <li><a href="calculator.html">🧮 GPA Calculator</a></li>
-            <li><a href="simulator.html">🎯 Goal Simulator</a></li>
-            <li><a href="mark_scaling.html">📈 Mark Scaling</a></li>
-            <li><a href="transcript.html">📜 Transcript</a></li>
-            <li><a href="settings.html">👤 Profile Settings</a></li>
-            <li class="mt-auto"><a href="#" id="logoutBtn" style="color: #f87171;">🚪 Logout</a></li>
-        </ul>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+    />
+    <link rel="stylesheet" href="../style/style.css" />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600&display=swap"
+      rel="stylesheet"
+    />
+  </head>
+  <body>
+    <div class="sidebar">
+      <div class="sidebar-header">🎓 Success Hub</div>
+      <ul class="sidebar-menu">
+        <li><a href="dashboard.html">📊 Dashboard</a></li>
+        <li><a href="semesters.html" class="active">📅 Semesters</a></li>
+        <li><a href="courses.html">📚 Courses</a></li>
+        <li><a href="calculator.html">🧮 GPA Calculator</a></li>
+        <li><a href="simulator.html">🎯 Goal Simulator</a></li>
+        <li><a href="mark_scaling.html">📈 Mark Scaling</a></li>
+        <li><a href="transcript.html">📜 Transcript</a></li>
+        <li><a href="settings.html">👤 Profile Settings</a></li>
+        <li class="mt-auto">
+          <a href="#" id="logoutBtn" style="color: #f87171">🚪 Logout</a>
+        </li>
+      </ul>
     </div>
 
-    <button id="mobile-toggle" class="mobile-toggle" aria-label="Toggle Menu">
-        ☰
-    </button>
-
     <div class="main-content">
-        <div class="d-flex justify-content-between align-items-center mb-4">
-            <h2 class="fw-bold">My Semesters</h2>
-            <button id="addSemesterBtn" class="btn btn-primary">+ Add New Semester</button>
-        </div>
+      <div class="d-flex justify-content-between align-items-center mb-4">
+        <h2 class="fw-bold">My Semesters</h2>
+        <button id="addSemesterBtn" class="btn btn-primary">
+          + Add New Semester
+        </button>
+      </div>
 
-        <div class="row">
-            <div class="col-12">
-                <div class="custom-card">
-                    <div class="table-responsive">
-                        <table class="table table-hover align-middle">
-                            <thead class="table-light">
-                                <tr>
-                                    <th>Semester Name</th>
-                                    <th>Date/Period</th>
-                                    <th class="text-end">Actions</th>
-                                </tr>
-                            </thead>
-                            <tbody id="semester-list">
-                                <!-- Injected by JS -->
-                            </tbody>
-                        </table>
-                    </div>
-                </div>
+      <div class="row">
+        <div class="col-12">
+          <div class="custom-card">
+            <div class="table-responsive">
+              <table class="table table-hover align-middle">
+                <thead class="table-light">
+                  <tr>
+                    <th>Semester Name</th>
+                    <th>Date/Period</th>
+                    <th class="text-end">Actions</th>
+                  </tr>
+                </thead>
+                <tbody id="semester-list">
+                  <!-- Injected by JS -->
+                </tbody>
+              </table>
             </div>
+          </div>
         </div>
+      </div>
     </div>
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     <script src="../scripts/ui.js"></script>
     <script src="../scripts/api.js" type="module"></script>
     <script type="module">
-        import { api } from '../scripts/api.js';
+      import { api } from "../scripts/api.js";
 
-        async function loadSemesters() {
-            try {
-                const semesters = await api.semesters.getAll();
-                const list = document.getElementById('semester-list');
-                list.innerHTML = "";
+      async function loadSemesters() {
+        try {
+          const semesters = await api.semesters.getAll();
+          const list = document.getElementById("semester-list");
+          list.innerHTML = "";
 
-                semesters.forEach(sem => {
-                    list.innerHTML += `
+          semesters.forEach((sem) => {
+            list.innerHTML += `
                         <tr>
                             <td class="fw-bold">${sem.name}</td>
-                            <td>${sem.start_date || 'Not specified'}</td>
+                            <td>${sem.start_date || "Not specified"}</td>
                             <td class="text-end">
                                 <a href="semester_view.html?id=${sem.id}" class="btn btn-sm btn-info text-white">View Details</a>
                                 <button class="btn btn-sm btn-outline-secondary" onclick="editSemester(${sem.id})">Edit</button>
@@ -81,31 +86,33 @@
                             </td>
                         </tr>
                     `;
-                });
-            } catch (err) {
-                window.location.href = 'login.html';
-            }
+          });
+        } catch (err) {
+          window.location.href = "login.html";
         }
+      }
 
-        window.deleteSemester = async (id) => {
-            if (confirm('Are you sure?')) {
-                try {
-                    await api.semesters.delete(id);
-                    loadSemesters();
-                } catch (err) { alert(err.message); }
-            }
-        };
+      window.deleteSemester = async (id) => {
+        if (confirm("Are you sure?")) {
+          try {
+            await api.semesters.delete(id);
+            loadSemesters();
+          } catch (err) {
+            alert(err.message);
+          }
+        }
+      };
 
-        window.editSemester = (id) => {
-            window.location.href = `semester_manage.html?id=${id}`;
-        };
+      window.editSemester = (id) => {
+        window.location.href = `semester_manage.html?id=${id}`;
+      };
 
-        document.getElementById('addSemesterBtn').onclick = () => {
-            window.location.href = 'semester_manage.html';
-        };
+      document.getElementById("addSemesterBtn").onclick = () => {
+        window.location.href = "semester_manage.html";
+      };
 
-        window.onload = loadSemesters;
+      window.onload = loadSemesters;
     </script>
     <script src="../scripts/theme.js"></script>
-</body>
+  </body>
 </html>
