+++ server/routes/ai.php
@@ -19,7 +19,7 @@ if ($method === 'POST') {
     }
 
     // Gemini API Endpoint
-    $url = "https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent?key=" . $apiKey;
+    $url = "https://generativelanguage.googleapis.com/v1beta/models/gemini-2.5-flash:generateContent?key=" . $apiKey;
 
     $payload = [
         "contents" => [
